+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/utils/message/MessageUtils.kt
@@ -97,8 +97,8 @@ public inline fun Message.isPinned(
 /**
  * @return If the message is a valid pinned message.
  */
-@PublishedApi
-internal inline fun Message.isPinExpired(now: () -> Long): Boolean = pinExpires?.let { it.time > now() } ?: false
+@InternalStreamChatApi
+public inline fun Message.isPinExpired(now: () -> Long): Boolean = pinExpires?.let { it.time < now() } ?: false
 
 /**
  * @return If the message type is regular.