+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/offline/plugin/factory/StreamOfflinePluginFactory.kt
@@ -60,7 +60,10 @@ import kotlin.reflect.KClass
  *
  * @param appContext [Context]
  */
-public class StreamOfflinePluginFactory(private val appContext: Context) : PluginFactory, RepositoryFactory.Provider {
+public class StreamOfflinePluginFactory @JvmOverloads constructor(
+    private val appContext: Context,
+    private val now: () -> Long = { System.currentTimeMillis() },
+) : PluginFactory, RepositoryFactory.Provider {
 
     private val logger by taggedLogger("Chat:OfflinePluginFactory")
 
@@ -77,6 +80,7 @@ public class StreamOfflinePluginFactory(private val appContext: Context) : Plugi
             database = createDatabase(appContext, user),
             currentUser = user,
             scope = ChatClient.instance().inheritScope { SupervisorJob(it) },
+            now = now,
         )
     }