+++ stream-chat-android-client/src/test/java/io/getstream/chat/android/client/attachment/AttachmentUploaderTests.kt
@@ -26,16 +26,13 @@ import io.getstream.chat.android.positiveRandomInt
 import io.getstream.chat.android.randomFile
 import io.getstream.chat.android.randomString
 import io.getstream.chat.android.test.TestCall
-import io.getstream.result.Error
 import io.getstream.result.Result
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.runTest
 import org.amshove.kluent.shouldBeEqualTo
 import org.amshove.kluent.shouldBeInstanceOf
-import org.amshove.kluent.shouldBeNull
 import org.amshove.kluent.shouldNotBeNull
 import org.junit.Before
-import org.junit.Ignore
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.mockito.kotlin.any
@@ -59,30 +56,6 @@ internal class AttachmentUploaderTests {
     fun setup() {
         Shadows.shadowOf(MimeTypeMap.getSingleton())
             .addExtensionMimeTypMapping("jpg", "image/jpeg")
-    }
-
-    @Test
-    @Ignore("Current logic doesn't work so. Need to rewrite test")
-    fun `Should return attachment with properly filled data when sending file has failed`(): Unit = runTest {
-        val error = Error.GenericError(message = "")
-        val attachment = randomAttachments(size = 1).first()
-
-        val sut = Fixture()
-            .givenMockedFileUploads(channelType, channelId, Result.Failure(error))
-            .get()
-
-        val result = sut.uploadAttachment(channelType, channelId, attachment) as Result.Success
-
-        with(result.value) {
-            name shouldBeEqualTo attachment.upload!!.name
-            fileSize.shouldNotBeNull()
-            type.shouldNotBeNull()
-            mimeType.shouldNotBeNull()
-            uploadState shouldBeEqualTo Attachment.UploadState.Failed(error)
-            url.shouldBeNull()
-            imageUrl.shouldBeNull()
-            assetUrl.shouldBeNull()
-        }
     }
 
     @Test
@@ -147,7 +120,6 @@ internal class AttachmentUploaderTests {
             val url = attachment.upload!!.absolutePath
             val expectedAttachment = attachment.copy(
                 assetUrl = url,
-                url = url,
                 type = "file",
                 mimeType = "",
                 name = attachment.upload!!.name,
@@ -173,7 +145,6 @@ internal class AttachmentUploaderTests {
             val url = attachment.upload!!.absolutePath
             val expectedAttachment = attachment.copy(
                 assetUrl = url,
-                url = url,
                 type = "file",
                 mimeType = "",
                 name = attachment.upload!!.name,