+++ stream-chat-android-client/src/test/java/io/getstream/chat/android/client/parser2/testdata/AttachmentDtoTestData.kt
@@ -39,7 +39,6 @@ internal object AttachmentDtoTestData {
           "title": "title",
           "title_link": "titleLink",
           "type": "type",
-          "url": "url",
           "original_height": 100,
           "original_width": 100,
           "draft": true
@@ -61,7 +60,6 @@ internal object AttachmentDtoTestData {
         title = "title",
         title_link = "titleLink",
         type = "type",
-        url = "url",
         original_height = 100,
         original_width = 100,
         extraData = mapOf("draft" to true),
@@ -89,7 +87,6 @@ internal object AttachmentDtoTestData {
         title = null,
         title_link = null,
         type = null,
-        url = null,
         original_width = null,
         original_height = null,
         extraData = emptyMap(),