+++ CHANGELOG.md
@@ -105,6 +105,7 @@
 ## stream-chat-android-compose
 ### 🐞 Fixed
 - Fixed quoted message styling. [#5316](https://github.com/GetStream/stream-chat-android/pull/5316)
+- Fixed checkbox visibility in `MessageComposer`'s footer when dark theme is used. [#5318](https://github.com/GetStream/stream-chat-android/pull/5318)
 
 ### ⬆️ Improved
 - Sent messages from the MessageComposer are marked as read. [#5322](https://github.com/GetStream/stream-chat-android/pull/5322)
@@ -159,7 +160,6 @@
 ### 🐞 Fixed
 - Fixed deleted pinned messages being highlighted as pinned. [#5315](https://github.com/GetStream/stream-chat-android/pull/5315)
 - Fixed the url used when click on the link. [#5314](https://github.com/GetStream/stream-chat-android/pull/5314)
-- Fixed checkbox visibility in `MessageComposer`'s footer when dark theme is used. [#5318](https://github.com/GetStream/stream-chat-android/pull/5318)
 
 ### ⬆️ Improved
 - Enabled Strong Skipping Mode for Compose compiler and improved Compose performance. [#5303](https://github.com/GetStream/stream-chat-android/pull/5303)